% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{data.table.sortAnnotations}
\alias{data.table.sortAnnotations}
\title{Sorted Annotations}
\usage{
data.table.sortAnnotations(granges, chromosomes = NULL, verbose = FALSE)
}
\arguments{
\item{granges}{A data.table object containing CpG annotations. Should contain
'Name', 'chr' and 'pos' columns.}

\item{chromosomes}{A list of chromosomes which should be returned from the
full annotations. Returns all chromosomes if NULL.}

\item{verbose}{Should workflow messages be printed? (default: FALSE)}
}
\description{
Sorts annotations by position and chromosome.
}
