% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{extractSortedAnnotations}
\alias{extractSortedAnnotations}
\title{Sorted annotations}
\usage{
extractSortedAnnotations(rgset, chromosomes = NULL)
}
\arguments{
\item{rgset}{a _RGChannelSet_ or _RGChannelSetExtended_ object}

\item{chromosomes}{a list of chromosomes for which to extract annotations. Will
return all chromosomes if not specified.}
}
\description{
Get sorted annotations from _RGChannelSet_ or _RGChannelSetExtended_
object from specified chromosomes.
}
