% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{findMethRegions}
\alias{findMethRegions}
\title{Methylation Regions}
\usage{
findMethRegions(
  beta,
  granges,
  chromosomes = NULL,
  maxSplit = 50000,
  thresholdDistance = 5e+07,
  rateFactor = 10,
  scale = TRUE,
  minClusterSize = 5,
  compressionRatio = 10,
  seed = 23985034,
  allowParallel = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{beta}{Beta value (or m-value) matrix with probes in rows and
samples in columns. Row names should correspond to names of the annotation
(see \code{granges} argument).}

\item{granges}{CpG annotations. It is a GRanges object containing sorted
coordinates of all CpGs of interest. The names of this object should
correspond to CpG names and match row names of the beta value matrix.}

\item{chromosomes}{A list of chromosomes which should be returned from the
full annotations. Returns all chromosomes if NULL.}

\item{maxSplit}{Maximum split size, integer (default: 50000).}

\item{thresholdDistance}{Threshold distance}

\item{rateFactor}{Rate factor}

\item{scale}{Should beta value matrix be scaled? (default: TRUE)}

\item{minClusterSize}{Minimum number of CpGs per cluster, integer
(default: 5).}

\item{compressionRatio}{Compression ratio}

\item{seed}{A seed for random number generator.}

\item{allowParallel}{Should parallel processing be used. If you wish
to use multiple cores, register cores using \code{doParallel} package.}

\item{verbose}{Should workflow messages be printed? (default: FALSE)}
}
\description{
Methylation Regions
}
