% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{getValues}
\alias{getValues}
\title{Extract Beta or M-values}
\usage{
getValues(set, type = c("beta", "mvalues"))
}
\arguments{
\item{type}{what type of values should be used (_beta_ or _mvalues_)}

\item{rgset}{a _RGChannelSet_, _RGChannelSetExtended_ or _MethylSet_ object.}
}
\description{
Get beta or mvalues from _RGChannelSet_, _RGChannelSetExtended_ or _MethylSet_ object.
}
