% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{methRegions}
\alias{methRegions}
\title{Methylation Regions}
\usage{
methRegions(
  beta,
  granges,
  thresholdDistance = 5e+07,
  rateFactor = 10,
  scale = TRUE,
  minClusterSize = 5,
  compressionRatio = 10,
  seed = 23985034,
  verbose = FALSE
)
}
\arguments{
\item{beta}{Beta value (or m-value) matrix with probes in rows and
samples in columns. Row names should correspond to names of the annotation
(see \code{granges} argument).}

\item{granges}{CpG annotations. It is a GRanges object containing sorted
coordinates of all CpGs of interest. The names of this object should
correspond to CpG names and match row names of the beta value matrix.}

\item{thresholdDistance}{Threshold distance}

\item{rateFactor}{Rate factor}

\item{scale}{Should beta value matrix be scaled? (default: TRUE)}

\item{minClusterSize}{Minimum number of CpGs per cluster, integer
(default: 5).}

\item{compressionRatio}{Compression ratio}

\item{seed}{A seed for random number generator.}

\item{verbose}{Should workflow messages be printed? (default: FALSE)}
}
\value{
Returns a named list. The names correspond to CpG names and the
values are for that CpG assigned cluster. Note that not all CpGs that
were in the original dataset will be returned. Some may not cluster at
all, and thus will not be returned.
}
\description{
Calculates methylation regions for a subset of the dataset.
When using this method, a single chromosome or its subset should be passed.
}
