% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmrPlot.R
\name{plotDMR}
\alias{plotDMR}
\title{DMR plot}
\usage{
plotDMR(beta, cpgs, annotation, conditions)
}
\arguments{
\item{beta}{A beta value matrix.}

\item{cpgs}{A list of CpGs to visualize.}

\item{annotation}{CpG annotations containing position coordinates and CpG names.}

\item{conditions}{A named list of conditions. The values should contain sample groups
while the names should be sample IDs (same as colnames of the beta value matrix).}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots methylation region. Displays beta values for each same and
draws a curve across the mean group methylation value of each position.
}
