% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{prepareBeta}
\alias{prepareBeta}
\title{Beta Values}
\usage{
prepareBeta(beta, scale = TRUE, verbose = TRUE, logFunction = message)
}
\arguments{
\item{beta}{Beta value matrix.}

\item{scale}{Should beta value matrix be scaled? (default: TRUE)}

\item{verbose}{Should workflow messages be printed? (default: FALSE)}

\item{logFunction}{Function to use for printing messages. Useful if you
wish to setup custom logging (default: message).}
}
\description{
Beta value preprocessing.
}
