% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{sortAnnotations}
\alias{sortAnnotations}
\title{Sorted Annotations}
\usage{
sortAnnotations(granges, chromosomes = NULL, verbose = FALSE)
}
\arguments{
\item{granges}{An object of GRanges or data.frame type. For more details
please check ?GRanges.sortAnnotations and ?data.table.sortAnnotations.}

\item{chromosomes}{A list of chromosomes which should be returned from the
full annotations. Returns all chromosomes if NULL.}

\item{verbose}{Should workflow messages be printed? (default: FALSE)}
}
\description{
Sorts annotations by position and chromosome.
}
