% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{split}
\alias{split}
\title{GRanges Splitting}
\usage{
split(granges, maxSplit = 50000)
}
\arguments{
\item{granges}{GRanges object}

\item{maxSplit}{maximum split size, integer}
}
\description{
Split GRanges of a single chromosome into splits smaller than maxSplit size.
Takes into account the distance between individual CpGs and performs splitting at
locations which are the furthest from each other.
}
