% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{splitChromosomes}
\alias{splitChromosomes}
\title{GRanges Splitting}
\usage{
splitChromosomes(granges, maxSplit = 50000)
}
\arguments{
\item{granges}{GRanges object.}

\item{maxSplit}{Maximum split size, integer (default: 50000).}
}
\description{
Splits GRanges object into splits smaller than maxSplit size.
Each chromosome is split separately, between CpGs that are furthest apart.
}
