% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{data.table.sortAnnotations}
\alias{data.table.sortAnnotations}
\title{Sorted Annotations}
\usage{
data.table.sortAnnotations(
  granges,
  chromosomes = NULL,
  verbose = FALSE,
  logFunction = message
)
}
\arguments{
\item{granges}{A data.table object containing CpG annotations. Should contain
'Name', 'chr' and 'pos' columns.}

\item{chromosomes}{A list of chromosomes which should be returned from the
full annotations. Returns all chromosomes if NULL.}

\item{verbose}{Should workflow messages be printed? (default: FALSE)}

\item{logFunction}{Function to use for printing messages. Useful if you
wish to setup custom logging (default: message).}
}
\description{
Sorts annotations by position and chromosome.
}
