% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{GRanges.sortAnnotations}
\alias{GRanges.sortAnnotations}
\title{Sorted Annotations}
\usage{
GRanges.sortAnnotations(granges, chromosomes = NULL, verbose = FALSE)
}
\arguments{
\item{granges}{A GRanges object. Should contain CpG IDs as names or have a
"Name" metadata column.}

\item{chromosomes}{A list of chromosomes which should be returned from the
full annotations. Returns all chromosomes if NULL.}

\item{verbose}{Should workflow messages be printed? (default: FALSE)}
}
\description{
Sorts annotations by position and chromosome.
}
