% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{methRegions}
\alias{methRegions}
\title{TODO: explain all variables
TODO: create an explanatory page with examples how to run the package
TODO: check all calculations
TODO: is there a way to optimize?
TODO: what could go wrong? at what scenarios will the sigmoid fail?
TODO: should this method be exported?
TODO: can it be used with m-values, not beta values? why? explanations!}
\usage{
methRegions(
  beta,
  granges,
  thresholdDistance = 5e+07,
  rateFactor = 10,
  scale = TRUE,
  minClusterSize = 5,
  compressionRatio = 10,
  seed = 23985034,
  verbose = FALSE
)
}
\description{
TODO: explain all variables
TODO: create an explanatory page with examples how to run the package
TODO: check all calculations
TODO: is there a way to optimize?
TODO: what could go wrong? at what scenarios will the sigmoid fail?
TODO: should this method be exported?
TODO: can it be used with m-values, not beta values? why? explanations!
}
