% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{split}
\alias{split}
\title{GRanges Splitting}
\usage{
split(granges, maxSplit = 50000)
}
\arguments{
\item{granges}{GRanges object.}

\item{maxSplit}{Maximum split size, integer (default: 50000).}
}
\description{
Splits GRanges of a single chromosome into splits smaller than maxSplit size.
Takes into account the distance between individual CpGs and splits the dataset
between CpGs that are furthest apart.
}
